Make partners its own script
This way it can be loaded more reliably.

The way we were including it before, there was no way to be 100%
sure that YUI had got around to loading core.loadPartners by the
time it was called.

So I've separated it out into its own YUI block, and instead of
calling it as a function I'm defining global variables (evil I
know, blame YUI) so that only when the partners function loads
will it be run.