Remove old India homepage
We used to redirect people to `/index_india` if they were in India
so we could show them an India-specific takeover. We then stopped
this in October 2015 (see #400). So I'm not removing the old
page (if you need it look in history), and redirecting
`/index_india` to `/` (the real homepage).