New style module, `.instructions`
The styling should be similar to the `.instruction` module,
but it works on a simple ordered list without extra `<div>` elements.

I'm intending to use this on the "How to create" pages in the download
section, so I haven't implemented an equivalent of the `instruction__details`
part yet, as I don't need it.