Separate get methods
Actually having an if to decide whether a feed should be json or rss is a bit ass-backwards.
It's not unreasonable to expect the user to know whether their feed is JSON or RSS.
So let's make two template tags: get_json_feed and get_rss_feed.
I've also renamed their corresponding functions with "_content" on the end, just to differentiate them.

Also, I've moved more of the functionality inside the original "get_*_content" functions,
as that seems to make sense.