Use new docker image
This will do a better job of detecting whether a database is needed.

I was having a problem with the old method - for some reason my django
container thought there was a service available called `db`, so it was
trying to connect to the database, and it was breaking Django. This
solves that.