Add two custom HTTP headers
'X-commit-ID' will display the value of the 'COMMIT_ID' variable, if it
exists. This variable should be included in the images that are built by
our new deployment system.

'X-k8s-pod' will display the value of the 'K8S_POD_NAME' variable, if it
exists. This variable should be passed to containers running in our
kubernetes deployments.