Handle missing 'searchInformation'
Fixes #3879, which mentioned errors in search. Looking in the logs,
I'm pretty sure the error was this:

> Internal Server Error: /search
> Traceback (most recent call last):
>   File "/usr/local/lib/python2.7/dist-packages/django/core/handlers/exception.py", line 41, in inner
>     response = get_response(request)
>   File "/usr/local/lib/python2.7/dist-packages/django/core/handlers/base.py", line 249, in _legacy_get_response
>     response = self._get_response(request)
>   File "/usr/local/lib/python2.7/dist-packages/django/core/handlers/base.py", line 187, in _get_response
>     response = self.process_exception_by_middleware(e, request)
>   File "/usr/local/lib/python2.7/dist-packages/django/core/handlers/base.py", line 185, in _get_response
>     response = wrapped_callback(request, *callback_args, **callback_kwargs)
>   File "/srv/webapp-12/code/c920db4/webapp/views.py", line 66, in search
>     results['searchInformation']['totalResults']
> KeyError: 'searchInformation'

It appears that occasionally, the API payload returned from Google
Custom Search API was simply missing the searchInformation key. There
doesn't seem to be any obvious pattern to when this key is missing
(it doesn't correlate with there being no results),
and it isn't mentioned in the API docs:

https://developers.google.com/custom-search/json-api/v1/reference/cse/list

Here I'm trying my best to continue even if this key is missing:

- If the key is missing, set "estimatedTotal" to None
- Only show the "Sorry we couldn't find" message if the estimatedTotal is *explicitly* zero, because otherwise there *may* be results

This should mean the search results page will work as normal most of the time -
the only slightly odd behaviour will be when there are no search results
and "searchInformation" is missing - in this case the user will see an
empty results set instead of the "Sorry we couldn't find" message.