Expose the error message in search results
Since the Google Custom Search API is clearly a bit unreliable in what
it returns (see 339639a and #3879), when it causes search to break it
would be helpful to know what went wrong without digging through logs.

To that end, here I am exposing the error message on the error page,
for search retrieval errors.