Point "Engage" heading at /resources
The "Engage" link points to /engage, which redirects to /resources.

This causes linkchecker to break in the jenkins jobs, and it's also
needlessly inefficient.

Since /engage doesn't actually exist, it would be better if the "Engage" breadcrumb nav "heading" simply wasn't a link - didn't go anywhere. But supporting that is more complex.