Change how languages are specified in takeovers
This changes how languages should be specified in takeovers.

Takeovers that are only for a specific language parent group,
the "lang" attribute should be specified. For takeovers that
are for all language parent groups, there should be no "lang"
attribute.

Currently there is no support for multiple languages, as we've never
needed it so far - if we want to implement it in future, I recommend
that we look for a new attribute called `data-lang-extra` that would
include extra languages for matching.