Use a higher z-index for the navigation
Since the navigation needs to overlay practically everything,
it should be high enough to provide plenty of lower layers
for other things on the page that need to use z-index.

I'm following a vague standard, which may be enshrined in team
practices soon, that the page gets to use z-index in the range `0-50`,
`51-100` should be reserved for plugins (like global-nav), and then
`100+` should be used to override plugins if needed.