Fix extends_with_args
Okay this was a difficult one to debug - we were getting an error
about 1 in 5 or 6 times of loading blog articles. It looks like this
was an obscure sort of race-condition in the template renderer:

It seems that ExtendsWithArgsNode.render() was being called multiple
times in the object's lifetime, for some reason - maybe multithreading
requests?

This mutated the `self.kwargs` array so that non-existent template
variables (e.g. article.title.rendered) were replaced with
"" (empty string), and then in the next run,
`template.Variable(value).resolve(context)` would be run against that
empty string.

So I've rewritten extends_with_args and ExtendsWithArgsNode
to make it generally tidier, but also to make it so its state
doesn't get manipulated by calling `.render()`. I've also
made it check for empty strings and handle them safely.