Move Talisker to top of dependencies
Talisker is quite fussy about its versions, and has caused a
number of bugs in the past. Currently I'm trying to fix this error
we're seeing in Sentry (although I'm not sure if it's having any
adverse impact):
https://sentry.is.canonical.com/canonical/ubuntu-com/issues/2172/

AttributeErrortalisker.prometheus in write_metrics
errormodule 'prometheus_client.core' has no attribute '_MmapedDict'
talisker/prometheus.py in write_metrics at line 349

Pip is not very clever about resolving versions for different packages,
and it will just install sub-dependencies which match the requirements
of the first package it encounters. In this case, this meant that
the first package it encountered was `canonicalwebteam.get-feeds`,
whose requirement for prometheus_client is ">0.1.1". So it thinks
"fine let's install the latest version, 0.7.1". It doesn't care that
Talisker, further down in the dependency list, requires "<0.5.0".

By moving Talisker to the top, it means that `pip` will choose
versions of dependencies that conform to Talisker's requirements
first, and only then process other modules - in this case
leading to us having prometheus_client 0.4.2, instead of 0.7.1.

Hopefully this might fix the error we're seeing in Sentry, although
we won't know until we release it as I don't know how to reproduce
it locally.

But even if it doesn't fix it, this seems like a wiser order for our
dependencies in general.