Make logout return_to path more sensible
Here I've changed the logic of the `/logout` view, so it also can accept a `return_to` paramter, and so the "Sign out" link on `/advantage` now passes a `return_to` parameter to bring you back to `/advantage`.

This is because in general I don't think it's sensible for `/logout` to send you back to the page you came from. In many cases the page you came from would only be available with authentication, and so you'd end up in a redirect loop.

If you want the user to come back to the existing page, you should explicitly set `return_to`, otherwise the user will be sent to the homepage, which I think is sensible.