Support old hash links
This supports #0, #1 style hashes in the tutorials, by
redirecting them to the appropriate section.

I've also changed the format of the section IDs to {num}-{description}
to avoid clashin with IDs in the global nav, and I've updated the global
nav JavaScript to handle IDs it doesn't recognise more gracefully.