Security API authentication proof of concept
This adds an "authorization_required" decorator for securing API
endpoints for the security section, and applies it to a
POST endpoint at /security/notices.

This will expect to be provided with a valid macaroonbakery macaroon,
and if it doesn't find one, will return a macaroonbakery JSON data
structure to ask the client to forward the user to
api.jujucharms.com/identity to login.

You can test this out with a script like this one:
https://gist.githubusercontent.com/nottrobin/8e89667066815a263021a5d3ba248e22/raw/10624702996ad4e68a5bd0cbc731b237df726691/post-cve.py