Use Talisker session for docs calls
Instead of the CachedSession from canonicalwebteam.http.

This is because we're getting timeout errors in Sentry,
because the CachedSession has a timeout of (.5, 3) by default.

We should no longer explicitly time out API calls, as
the combination of the content-cache in front of our sites
and the fact that we use gsync gevent workers in production
should mean that there's no problem with request waiting for
the API calls to complete.

Also, using a Talisker session allows us to get more information
about these API requests in Graylog.