Replace Wordpress class with BlogAPI class
`BlogAPI` class acts like a wrapper for the `Wordpress` class.
It will always do the article transform, and it accepts a boolean
parameter to decide if we want url transformation or not.

Url tranformation replaces `admin.insights.ubuntu.com/wp-content/uploads`
with `ubuntu.com/wp-content/uploads`. Additionally it enables the cloudinary
service for our images again, which was turned off for unknown reasons.