Add information tool-tip for CVE statuses
Add tool-tip to give details on CVE status and priority.

We have no texts for the following statuses:
needs triage, needed, deferred, ignored, released, DNE

Right now cells with those statuses will show no status explanation in the tool-tip.

`Not vulnerable` and `Does not exist` do not show any priority explanation in the tool-tip because it does not show a priority icon in the cell.

`Does not exist` does not display the information icon since it's missing status explanation and has no priority icon