Rework MicroStack docs to use Secure Clustering
As of revision 215 in the store, MicroStack contains the Secure
Clustering feature:
https://review.opendev.org/#/c/757658/

The workflow no longer includes a cluster password and instead relies on
a connection string (more details in the opendev review).

It is also mandatory to specify the node role via command line when
using the --auto mode too.

Passwords are now randomly generated to avoid insecure practices which
is also reflected in the documentation:
https://review.opendev.org/#/c/756567/