Support git+ protocol in requirements.txt
We often want to demo a change from a module branch, and the simplest
way to do it is to include a line in requirements.txt like
`git+git://github.com/owner/repo@branch#module`

This was failing in demos because the docker image doesn't install git,
which is needed.

This will optionally install git if it notices that the requirements.txt
includes a git+ line.