Create FinancialServices product card
Created a new product card for https://ubuntu.com/engage/finance

Image: was uploaded by me on the assets. If needed I can provide it. If possible, make it bigger in the product card, as it is the main messaging around the ads
Blog post tag: Events (this is the tag where it should be shown)