Fixed bug with `p-code-copyable`
I investigated a bit, because `main.js` is already present on my lsn template page, and as soon as I delete the block of javascript, the copy does not work anymore.

I have detected the usage of the `p-code-copyable` on these pages:
https://ubuntu.com/kubernetes/install
https://ubuntu.com/internet-of-things/edgex
https://ubuntu.com/ai/install-kubeflow
https://ubuntu.com/openstack/install

Out of these, the only one working is `/internet-of-things/edgex` because @anthonydillon  also added the script on the page.

`main.js` has some `p-code-copyable` related js inside `/static/scratch.js`, but it seems to be not working.

So I will take the liberty to to replace the code with the script from the docs: https://vanillaframework.io/docs/base/code#copyable
(also I will delete ant's copyable code from `/internet-of-things/edgex`)

As a result all the above links' copyable elements started to work.