Exclude x-frame-options header for templatefinder
The x-frame-options is primarily to protect us against clickjacking attacks against logged-in users.

(The attack would involve hiding an iframe of one of our sites behind a link or button on the attacker's site, and presenting that site to a user who is already logged in to our site. You could then trick the user into unknowingly clicking on a logged-in action on our site.)

However, the TemplateFinder pages are always static content, and so don't include any logged-in actions that could be exploited.

The x-frame-options header caused an issue where a PDF is supposed to include a frame of https://ubuntu.com/legal/ubuntu-advantage-service-terms. If we're too broad with x-frame-options we could well see further issues of the same type.