Update payment-methods page
Remove the /payment-methods link from the sub-nav.

If a user has no ubuntu one account and they try to go to
/payment-methods, we send them back to /advantage

If a user has an ubuntu one account but no stripe account we throw a
500.

Users should have no business on the /payment-methods page without both
a ubuntu one account or a stripe account.