Add primitives and models files
We are splitting our classes in two types:
- Primitives
- Models

This is so that we have a clear separation between
objects that belong to the domain of the API we consume and
the object that belong to the domain of Ubuntu Advantage.

Primitives are classes that are built from extracting the
necessary properties of the API json responses.

Examples of primitives:
Contract, ContractItem, Subscription, Product, Account

Models are classes that are built by processing primitive classes.
These are refined structures designed specifically for Ubuntu Advantage.

Examples of models: UserSubscription, Listings