Add builders, helpers and parsers files
We are splitting the functions depending on the action they fulfil.

Parsers - functions that accept raw json data and convert it
to a primitive or a list of primitives. Name ideally starting with `parse_`.

Builders - functions that build specific structures. Name ideally
starting with `build_`.

Helpers - functions that are not parsers or builders. They have a more
general purpose.