Fix _make_renewal tests
Adding the API object to the flask.g global object makes testing hard, as this can't be easily mocked.

Even though it's a lot more lines of code, I'm inclined to think it's better to declaratively instantiate the api object at the start of each view. It's a lot easier to see what's going on. The API object is one of the most important parts of the code, so it's good to make it clear how it's instantiated etc., so there's no confusion.

So here I remove the API object from flask.g (so we're no longer using flask.g anywhere) and instead instantiate the API object at the top of each view.

This then means that we can pass the API object directly through to the _make_renewal function, which then allows us to mock it in the test.