fix exception on hardware details page
Fixes #11403

On certificates with no hardware details the hardware
details through an exception due to an attempt to
use a variable that didn't get set.

On review, the variable wasn't actually needed.

I removed it, and also updated the hardware details
page to show a message when no details are available.