Fix IoT and Server SoC filter
This is regression due to the security patch. Check Ubuntu Core and Server SoC during parsing, so that page redirects with the correct params. Removed the bleach function, it's not longer needed since we are checking every param given by the user with the API data using _parse_query_params. So in this case bleach is causing more harm than good.