Update and replace Kernel release diagram (#13922)
The reason of removing the tabs is that every time we have a new update, we'll need to update in 2 different places,
with different sources of truth, plus the separation of tabs (TLS, separate new section). Since the table is not that big
I replaced it completely to use one single component and one single source of truth, so everytime there is an update, we'll just update it in one place. This sets the basis for the release.yaml project to avoids future "copy updates".